refactor: make the cdk_resource_invoke construct generic over lambda function type
