+++ lib/workload/stateless/functions/cdk_resource_invoke.ts
@@ -5,29 +5,60 @@ import {
   AwsSdkCall,
   PhysicalResourceId,
 } from 'aws-cdk-lib/custom-resources';
-import * as fn from './functions/function';
+import * as lambda from 'aws-cdk-lib/aws-lambda';
 import { CfnOutput, Stack, Token } from 'aws-cdk-lib';
 import { IVpc, SubnetType } from 'aws-cdk-lib/aws-ec2';
 import { ManagedPolicy, PolicyStatement, Role, ServicePrincipal } from 'aws-cdk-lib/aws-iam';
+import { Version } from 'aws-cdk-lib/aws-lambda';
+
+/**
+ * The interface by which the generic type of `CdkResourceInvoke` is constrained by.
+ * Note that the standard `lambda.Function` meets this constraint, which means it can
+ * be used directly with the `CdkResourceInvoke`, rather than having to wrap it in
+ * another class.
+ */
+export interface InvokeFunction {
+  /**
+   * The function name to be used when constructing the invoke function.
+   */
+  functionName?: string;
+  /**
+   * The current version of the function.
+   */
+  currentVersion: Version;
+}
+
+/**
+ * An interface representing the function name that is created when using `CdkResourceInvoke`.
+ */
+export type FunctionName = {
+  /**
+   * Function name.
+   */
+  functionName: string;
+}
 
 /**
  * Props for the resource invoke construct.
  */
-export type CdkResourceInvokeProps = {
+export type CdkResourceInvokeProps<P, F extends InvokeFunction> = {
   /**
    * Vpc for the function.
    */
   vpc: IVpc;
   /**
-   * The function to create. This will override the function name to ensure that it remains
+   * The function to create. This will override any `functionName` property to ensure that it remains
    * callable using the singleton function created by `AwsCustomResource`. See
    * https://github.com/aws-samples/amazon-rds-init-cdk/blob/239626632f399ebe4928410a49d5ac5d009a6502/lib/resource-initializer.ts#L69-L71.
+   *
+   * It is expected that this creates a Lambda function with the `functionName`. This allows `CdkResourceInvoke`
+   * to call the function after it is created.
    */
-  createFunction: (scope: Construct, id: string, props: fn.FunctionPropsNoPackage) => fn.Function;
+  createFunction: (scope: Construct, id: string, props: FunctionName & P) => F;
   /**
    * Function props when creating the Lambda function.
    */
-  functionProps: fn.FunctionPropsNoPackage;
+  functionProps: P;
   /**
    * Name to use when creating the function.
    */
@@ -39,14 +70,16 @@ export type CdkResourceInvokeProps = {
 };
 
 /**
- * A construct for invoking a Lambda function for resource initialization.
+ * A construct for invoking a Lambda function for resource initialization. This is useful for performing
+ * database actions such as migrations during CloudFormation stack creation, to ensure that a database is
+ * in the expected state before the stack succeeds.
  */
-export class CdkResourceInvoke extends Construct {
+export class CdkResourceInvoke<P, F extends InvokeFunction> extends Construct {
   private readonly _response: string;
   private readonly _customResource: AwsCustomResource;
-  private readonly _function: fn.Function;
+  private readonly _function: F;
 
-  constructor(scope: Construct, id: string, props: CdkResourceInvokeProps) {
+  constructor(scope: Construct, id: string, props: CdkResourceInvokeProps<P, F>) {
     super(scope, id);
 
     const stack = Stack.of(this);
@@ -60,10 +93,10 @@ export class CdkResourceInvoke extends Construct {
       service: 'Lambda',
       action: 'invoke',
       parameters: {
-        FunctionName: this.function.functionName(),
+        FunctionName: this.function.functionName,
       },
       physicalResourceId: PhysicalResourceId.of(
-        `${id}-AwsSdkCall-${this.function.currentVersion()}`
+        `${id}-AwsSdkCall-${this.function.currentVersion}`
       ),
     };
 
@@ -130,7 +163,7 @@ export class CdkResourceInvoke extends Construct {
   /**
    * Get the function.
    */
-  get function(): fn.Function {
+  get function(): F {
     return this._function;
   }
 }
